+++ multiplayer_course/3/example_ready_state.py
@@ -65,4 +65,4 @@ def multiplayer_main(net: s.NetClient, role: str) -> None:
 
 
 if __name__ == "__main__":
-    s.networking.run()
+    s.run(multiplayer=True, multiplayer_entry=multiplayer_main)